Adding a class to deal with accessing env variables, dto for provider config and a new class to manage allproviders config, change factory to use this class, and subsequencial classes. First draft of the dto configuration.

+++ src/Services/Credit/CreditProviderInterface.php
@@ -3,5 +3,5 @@
 namespace App\Services\Credit;
 interface CreditProviderInterface
 {
-    public function getRates(int $amount): array;
+    public function getRates(float $amount): array;
 }
